requirements.txt
ansible-core>=2.15,<2.18
molecule>=6.0,<7.0
pytest>=7.0,<9.0
pytest-testinfra>=10.0,<11.0
pyyaml>=6.0
